Huge shift in folder structure. Recovered changes from changing computer also didnt help with the size of the change.
