--- src/Containers/footer.js
@@ -1,14 +1,0 @@
-import React from "react";
-
-
-const Footer = () => {
-  return (
-    <div className="bottom-0 flex items-center justify-center w-full h-8 px-2 py-1 text-xs text-gray-500 ">
-      <a href="https://www.romainclemencon.com/" target="_blank">
-        &nbsp; Romain Clemencon 2023
-      </a>
-    </div>
-  );
-};
-
-export default Footer;
