+++ src/Pages/Exercise/outputDetails.js
