+++ src/index.css
@@ -4,6 +4,12 @@
 @tailwind components;
 @tailwind utilities;
 
+@layer components {
+  .exercise-sep {
+    @apply mx-auto my-4 h-1 w-[30rem] rounded border-0 bg-gray-300;
+  }
+}
+
 * {
   box-sizing: border-box;
   margin: 0;
@@ -49,12 +55,6 @@ nav > div > a.active {
   
   } 40%, 80% {
     transform:rotate(10deg);
-  }
-}
-
-@layer components {
-  .exercise-sep {
-    @apply mx-auto my-4 h-1 w-[30rem] rounded border-0 bg-gray-300;
   }
 }
