+++ src/index.js
@@ -4,7 +4,6 @@ import AuthProvider from "./Contexts/authContext";
 
 import "./index.css";
 import App from "./App";
-import Footer from "./Containers/footer";
 
 
 
@@ -15,6 +14,5 @@ ReactDOM.createRoot(document.getElementById("root")).render(
         <App />
       </AuthProvider>
     </React.StrictMode>
-    <Footer />
   </>
 );
